Refactor service common
