+++ auth-service/src/main/java/com/ercanbeyen/authservice/advice/AuthServiceGlobalExceptionHandler.java
@@ -3,8 +3,8 @@ package com.ercanbeyen.authservice.advice;
 import com.ercanbeyen.authservice.exception.InvalidUserCredentialException;
 import com.ercanbeyen.authservice.exception.TokenAlreadyRevokedException;
 import com.ercanbeyen.authservice.exception.UserAlreadyExistException;
-import com.ercanbeyen.servicecommon.client.exception.ErrorResponse;
-import com.ercanbeyen.servicecommon.client.exception.GlobalErrorCode;
+import com.ercanbeyen.servicecommon.client.exception.response.ErrorResponse;
+import com.ercanbeyen.servicecommon.client.exception.constant.GlobalErrorCode;
 import io.jsonwebtoken.ExpiredJwtException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
