+++ auth-service/src/main/java/com/ercanbeyen/authservice/config/CustomUserDetails.java
@@ -9,11 +9,10 @@ import java.util.Collection;
 import java.util.List;
 import java.util.stream.Collectors;
 
-
 public class CustomUserDetails implements UserDetails {
-    private String username;
-    private String password;
-    private List<GrantedAuthority> authorities;
+    private final String username;
+    private final String password;
+    private final List<GrantedAuthority> authorities;
 
     public CustomUserDetails(UserCredential userCredential) {
         this.username = userCredential.getUsername();
