--- auth-service/src/main/java/com/ercanbeyen/authservice/constant/enums/Role.java
@@ -1,6 +1,0 @@
-package com.ercanbeyen.authservice.constant.enums;
-
-public enum Role {
-    USER,
-    ADMIN
-}
