+++ school-service/src/main/java/com/ercanbeyen/schoolservice/advice/SchoolServiceGlobalExceptionHandler.java
@@ -1,6 +1,8 @@
 package com.ercanbeyen.schoolservice.advice;
 
 import com.ercanbeyen.servicecommon.client.exception.*;
+import com.ercanbeyen.servicecommon.client.exception.constant.GlobalErrorCode;
+import com.ercanbeyen.servicecommon.client.exception.response.ErrorResponse;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
