+++ service-common/src/main/java/com/ercanbeyen/servicecommon/client/logging/LogMessage.java
@@ -4,7 +4,7 @@ public class LogMessage {
     public static final String RESOURCE_FOUND = "{} {} is found";
     public static final String CLIENT_SERVICE_RESPONSE = "Client {} Service Response: {}";
     public static final String NOTIFICATION_SENT_TO_QUEUE = "Notification is sent to topic {} at {}";
-    public static final String LOGGED_IN_USER = "loggedInUsername: {}";
+    public static final String LOGGED_IN_USER = "logged in username: {}";
 
     private LogMessage() {}
 }
